requirements.txt
Flask==1.1.2
colorama==0.4.4
python-dotenv==0.18.0
mysql-connector-python==8.0.25